refactor(compiler-cli): move downlevel decorators transform into ngtsc
This commit moves the `downlevel_decorators_transform` into a dedicated
compilation unit within `ngtsc`.

The tests were refactored to move to ngtsc's testing abstractions, allowing
VE's `MockAotContext` and `MockCompilerHost` to be removed.
