+++ packages/compiler-cli/BUILD.bazel
@@ -76,6 +76,7 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/core:api",
         "//packages/compiler-cli/src/ngtsc/diagnostics",
         "//packages/compiler-cli/src/ngtsc/docs",
+        "//packages/compiler-cli/src/ngtsc/downlevel_decorators",
         "//packages/compiler-cli/src/ngtsc/file_system",
         "//packages/compiler-cli/src/ngtsc/incremental",
         "//packages/compiler-cli/src/ngtsc/indexer",
@@ -88,7 +89,6 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/typecheck",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
         "//packages/compiler-cli/src/ngtsc/util",
-        "//packages/compiler-cli/src/transformers/downlevel_decorators_transform",
         "@npm//@bazel/concatjs",
         "@npm//@types/node",
         "@npm//@types/yargs",
