+++ packages/compiler-cli/src/ngtsc/downlevel_decorators/BUILD.bazel
@@ -3,8 +3,10 @@ load("//tools:defaults.bzl", "ts_library")
 package(default_visibility = ["//visibility:public"])
 
 ts_library(
-    name = "downlevel_decorators_transform",
-    srcs = glob(["*.ts"]),
+    name = "downlevel_decorators",
+    srcs = glob([
+        "**/*.ts",
+    ]),
     deps = [
         "//packages/compiler-cli/src/ngtsc/imports",
         "//packages/compiler-cli/src/ngtsc/reflection",
