+++ packages/compiler-cli/src/ngtsc/downlevel_decorators/index.ts
@@ -6,4 +6,4 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export {getDownlevelDecoratorsTransform} from './downlevel_decorators_transform';
+export {getDownlevelDecoratorsTransform} from './src/downlevel_decorators_transform';
