+++ packages/compiler-cli/test/BUILD.bazel
@@ -6,7 +6,6 @@ ts_library(
     name = "test_utils",
     testonly = True,
     srcs = [
-        "mocks.ts",
         "test_support.ts",
     ],
     visibility = [
@@ -49,28 +48,6 @@ jasmine_node_test(
         "//packages/common:npm_package",
         "//packages/core",
         "@npm//yargs",
-    ],
-)
-
-ts_library(
-    name = "downlevel_decorator_transform_lib",
-    testonly = True,
-    srcs = [
-        "downlevel_decorators_transform_spec.ts",
-    ],
-    deps = [
-        ":test_utils",
-        "//packages/compiler-cli/src/ngtsc/reflection",
-        "//packages/compiler-cli/src/transformers/downlevel_decorators_transform",
-        "@npm//typescript",
-    ],
-)
-
-jasmine_node_test(
-    name = "downlevel_decorator_transform",
-    bootstrap = ["//tools/testing:node"],
-    deps = [
-        ":downlevel_decorator_transform_lib",
     ],
 )
